IRT-5151 Use Company object instead of TpmObjectMetadata. Init Profile (county code) for Company/Subsidiary. Refactor TradingPartnerClient.
