+++ src/main/java/com/figaf/integration/tpm/client/TpmBaseClient.java
@@ -6,9 +6,15 @@ import com.fasterxml.jackson.databind.SerializationFeature;
 import com.figaf.integration.common.client.BaseClient;
 import com.figaf.integration.common.factory.HttpClientsFactory;
 import com.figaf.integration.tpm.entity.AdministrativeData;
+import com.figaf.integration.tpm.entity.trading.verbose.*;
+import org.json.JSONArray;
 import org.json.JSONObject;
 
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
+
+import static com.figaf.integration.common.utils.Utils.optString;
 
 /**
  * @author Kostas Charalambous
@@ -60,9 +66,161 @@ public abstract class TpmBaseClient extends BaseClient {
         AdministrativeData administrativeData = new AdministrativeData();
         administrativeData.setCreatedAt(new Date(administrativeDataJsonObject.getLong("createdAt")));
         administrativeData.setModifiedAt(new Date(administrativeDataJsonObject.getLong("modifiedAt")));
-        administrativeData.setCreatedBy(administrativeDataJsonObject.getString("createdBy"));
-        administrativeData.setModifiedBy(administrativeDataJsonObject.getString("modifiedBy"));
+        administrativeData.setCreatedBy(administrativeDataJsonObject.optString("createdBy"));
+        administrativeData.setModifiedBy(administrativeDataJsonObject.optString("modifiedBy"));
         return administrativeData;
+    }
+
+    protected ArtifactProperties parseArtifactProperties(final JSONObject jsonArtifactProperties) {
+        ArtifactProperties artifactProperties = new ArtifactProperties();
+        artifactProperties.setWebURL(optString(jsonArtifactProperties, "webURL"));
+        artifactProperties.setShortName(optString(jsonArtifactProperties, "shortName"));
+        artifactProperties.setLogoID(optString(jsonArtifactProperties, "logoID"));
+        return artifactProperties;
+    }
+
+    protected ProfileDto parseProfileDto(JSONObject profileJsonObject) {
+        ProfileDto profile = new ProfileDto();
+
+        JSONObject jsonAddress = profileJsonObject.getJSONObject("Address");
+        profile.setAddress(parseAddress(jsonAddress));
+
+        JSONArray jsonContactPersonList = profileJsonObject.getJSONArray("ContactPersonList");
+        profile.setContactPersonList(parseContactPersons(jsonContactPersonList));
+
+        JSONObject jsonBusinessContext = profileJsonObject.getJSONObject("BusinessContext");
+        BusinessContext BusinessContext = new BusinessContext();
+
+        if (jsonBusinessContext.has("BusinessProcess")) {
+            JSONObject jsonBusinessProcess = jsonBusinessContext.getJSONObject("BusinessProcess");
+            BusinessContext.setBusinessProcess(parseBusinessProcessDTO(jsonBusinessProcess));
+        }
+
+        if (jsonBusinessContext.has("BusinessProcessRole")) {
+            JSONObject jsonBusinessProcessRole = jsonBusinessContext.getJSONObject("BusinessProcessRole");
+            BusinessContext.setBusinessProcessRole(parseBusinessProcessRoleDTO(jsonBusinessProcessRole));
+        }
+
+        if (jsonBusinessContext.has("IndustryClassification")) {
+            JSONObject jsonIndustryClassification = jsonBusinessContext.getJSONObject("IndustryClassification");
+            BusinessContext.setIndustryClassification(parseIndustryClassificationDTO(jsonIndustryClassification));
+        }
+
+        if (jsonBusinessContext.has("ProductClassification")) {
+            JSONObject jsonProductClassification = jsonBusinessContext.getJSONObject("ProductClassification");
+            BusinessContext.setProductClassification(parseProductClassificationDTO(jsonProductClassification));
+        }
+
+        if (jsonBusinessContext.has("GeoPolitical")) {
+            JSONObject jsonGeoPolitical = jsonBusinessContext.getJSONObject("GeoPolitical");
+            BusinessContext.setGeoPolitical(parseGeoPoliticalDTO(jsonGeoPolitical));
+        }
+
+        profile.setBusinessContext(BusinessContext);
+
+        return profile;
+    }
+
+    private BusinessProcess parseBusinessProcessDTO(JSONObject jsonBusinessProcess) {
+        BusinessProcess businessProcessDTO = new BusinessProcess();
+        businessProcessDTO.setBusinessProcessCodes(parseStringList(jsonBusinessProcess.optJSONArray("BusinessProcessCodeList")));
+        return businessProcessDTO;
+    }
+
+    private List<String> parseStringList(JSONArray jsonArray) {
+        List<String> stringList = new ArrayList<>();
+        if (jsonArray == null) {
+            return stringList;
+        }
+
+        for (int i = 0; i < jsonArray.length(); i++) {
+            stringList.add(jsonArray.optString(i));
+        }
+        return stringList;
+    }
+
+    private BusinessProcessRole parseBusinessProcessRoleDTO(JSONObject jsonBusinessProcessRole) {
+        BusinessProcessRole businessProcessRoleDTO = new BusinessProcessRole();
+        businessProcessRoleDTO.setBusinessProcessRoleCodes(parseStringList(jsonBusinessProcessRole.optJSONArray("BusinessProcessRoleCodeList")));
+        return businessProcessRoleDTO;
+    }
+
+    private IndustryClassification parseIndustryClassificationDTO(JSONObject jsonIndustryClassification) {
+        IndustryClassification industryClassificationDTO = new IndustryClassification();
+        industryClassificationDTO.setIndustryClassificationCodeList(parseStringList(jsonIndustryClassification.optJSONArray("IndustryClassificationCodeList")));
+        return industryClassificationDTO;
+    }
+
+    private ProductClassification parseProductClassificationDTO(JSONObject jsonProductClassification) {
+        ProductClassification productClassificationDTO = new ProductClassification();
+        productClassificationDTO.setProductClassificationCodeList(parseStringList(jsonProductClassification.optJSONArray("ProductClassificationCodeList")));
+        return productClassificationDTO;
+    }
+
+    private Address parseAddress(JSONObject jsonAddress) {
+        Address address = new Address();
+        address.setCityName(optString(jsonAddress, "CityName"));
+        address.setCountryCode(optString(jsonAddress, "CountryCode"));
+        address.setCountrySubdivisionCode(optString(jsonAddress, "CountrySubdivisionCode"));
+        address.setHouseNumber(optString(jsonAddress, "HouseNumber"));
+        address.setPOBox(optString(jsonAddress, "POBox"));
+        address.setPOBoxPostalCode(optString(jsonAddress, "POBoxPostalCode"));
+        address.setStreetName(optString(jsonAddress, "StreetName"));
+        address.setStreetPostalCode(optString(jsonAddress, "StreetPostalCode"));
+        return address;
+    }
+
+    private GeoPolitical parseGeoPoliticalDTO(JSONObject jsonGeoPolitical) {
+        GeoPolitical geoPoliticalDTO = new GeoPolitical();
+        JSONArray jsonCountryInfoList = jsonGeoPolitical.optJSONArray("CountryInfoList");
+        if (jsonCountryInfoList == null) {
+            return geoPoliticalDTO;
+        }
+        List<CountryInfo> countryInfos = new ArrayList<>();
+        for (int i = 0; i < jsonCountryInfoList.length(); i++) {
+            JSONObject jsonCountryInfo = jsonCountryInfoList.getJSONObject(i);
+            CountryInfo countryInfoDTO = new CountryInfo();
+            countryInfoDTO.setCountryCode(jsonCountryInfo.optString("CountryCode"));
+            countryInfoDTO.setCountrySubdivisionCode(jsonCountryInfo.optString("CountrySubdivisionCode"));
+            countryInfos.add(countryInfoDTO);
+        }
+        geoPoliticalDTO.setCountryInfoList(countryInfos);
+        return geoPoliticalDTO;
+    }
+
+    public List<ContactPersonDTO> parseContactPersons(JSONArray jsonContactPersonList) {
+        List<ContactPersonDTO> contactPersonDTOs = new ArrayList<>();
+        for (int i = 0; i < jsonContactPersonList.length(); i++) {
+            JSONObject jsonContactPerson = jsonContactPersonList.getJSONObject(i);
+            ContactPersonDTO contactPersonDTO = parseContactPerson(jsonContactPerson);
+            contactPersonDTOs.add(contactPersonDTO);
+        }
+        return contactPersonDTOs;
+    }
+
+    private ContactPersonDTO parseContactPerson(final JSONObject jsonContactPerson) {
+        ContactPersonDTO contactPersonDTO = new ContactPersonDTO();
+        contactPersonDTO.setId(optString(jsonContactPerson, "ID"));
+        contactPersonDTO.setFamilyName(optString(jsonContactPerson, "FamilyName"));
+        contactPersonDTO.setGivenName(optString(jsonContactPerson, "GivenName"));
+        contactPersonDTO.setPrimaryContact(jsonContactPerson.optBoolean("IsPrimaryContact"));
+        contactPersonDTO.setNewVersion(jsonContactPerson.optBoolean( "IsNewVersion"));
+        contactPersonDTO.setDocumentSchemaVersion(optString(jsonContactPerson, "DocumentSchemaVersion"));
+        contactPersonDTO.setArtifactStatus(optString(jsonContactPerson, "artifactStatus"));
+
+        if (jsonContactPerson.has("Address")) {
+            JSONObject jsonAddress = jsonContactPerson.getJSONObject("Address");
+            ContactPersonAddressDTO contactPersonAddressDTO = parseContactPersonAddress(jsonAddress);
+            contactPersonDTO.setAddress(contactPersonAddressDTO);
+        }
+
+        return contactPersonDTO;
+    }
+
+    private ContactPersonAddressDTO parseContactPersonAddress(final JSONObject jsonAddress) {
+        ContactPersonAddressDTO contactPersonAddressDTO = new ContactPersonAddressDTO();
+        contactPersonAddressDTO.setCountryCode(optString(jsonAddress, "CountryCode"));
+        return contactPersonAddressDTO;
     }
 
 }
